requirements.txt
asgiref==3.2.10
Django==3.1.2
django-datatables-view==1.19.1
factory-boy==3.1.0
Faker==4.5.0
install==1.3.4
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.0
text-unidecode==1.3
